refactor: partial migration of containers, block entities, entities, and models to 26.1 APIs
- Container: isClientSide field to isClientSide() method, recipe API, ItemStack Codec
- BlockEntity: GeckoLib AnimationController, ItemStack Codec, CustomData.copyTag()
- Entity: EntityReference owner, AnimationController 3-arg, snapTo, dropFromLootTable
- Model: BatFamiliarModel EntityRenderState, DefaultedJobEntityModel GeoRenderState
- GUI: mouseClicked/keyPressed/charTyped new event signatures
- ClientPentacleManager: recipe access refactor

+++ src/main/java/com/klikli_dev/occultism/common/entity/familiar/BatFamiliarEntity.java
@@ -121,7 +121,7 @@ public class BatFamiliarEntity extends FamiliarEntity implements FlyingAnimal {
     }
 
     @Override
-    public boolean causeFallDamage(float fallDistance, float damageMultiplier, DamageSource source) {
+    public boolean causeFallDamage(double fallDistance, float damageMultiplier, DamageSource source) {
         return false;
     }
